+++ previous-versions/assess/src/sdk/sdk.ts
@@ -55,8 +55,8 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "1.0";
-    sdkVersion = "0.34.0";
-    genVersion = "2.109.3";
+    sdkVersion = "0.34.1";
+    genVersion = "2.116.0";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);